requirements_app.txt
torch>=2.0
torchvision>=0.15
pillow>=9.0
pandas>=1.3
scikit-learn>=1.0